MovieScene Truncation Warning fixes
#jira UE-166271
#rb daren.cheng

[CL 31057708 by vincent gauthier in ue5-main branch]

+++ Engine/Source/Editor/SequencerCore/Public/TimeToPixel.h
@@ -24,7 +24,7 @@ public:
 		, TickResolution( InTickResolution )
 	{
 		ensureMsgf(WidthPx >= 0, TEXT("Got negative pixel width!"));
-		const double VisibleWidth = InLocalViewRange.Size<double>();
+		const float VisibleWidth = static_cast<float>(InLocalViewRange.Size<double>());
 
 		const float MaxPixelsPerSecond = 1000.f;
 		PixelsPerSecond = VisibleWidth > 0 ? WidthPx / VisibleWidth : MaxPixelsPerSecond;
@@ -38,7 +38,7 @@ public:
 	 */
 	float SecondsToPixel( double Time ) const
 	{
-		return (Time - ViewRangeStartSeconds) * PixelsPerSecond;
+		return static_cast<float>(Time - ViewRangeStartSeconds) * PixelsPerSecond;
 	}
 
 	/**
@@ -49,7 +49,7 @@ public:
 	 */
 	float SecondsDeltaToPixel( double TimeDelta ) const
 	{
-		return TimeDelta * PixelsPerSecond;
+		return static_cast<float>(TimeDelta) * PixelsPerSecond;
 	}
 
 	/**
@@ -71,7 +71,7 @@ public:
 	 */
 	float FrameToPixel( const FFrameTime& Time ) const
 	{
-		return (Time / TickResolution - ViewRangeStartSeconds) * PixelsPerSecond;
+		return static_cast<float>(Time / TickResolution - ViewRangeStartSeconds) * PixelsPerSecond;
 	}
 
 	/**
@@ -82,7 +82,7 @@ public:
 	 */
 	float FrameDeltaToPixel( const FFrameTime& TimeDelta) const
 	{
-		return (TimeDelta / TickResolution) * PixelsPerSecond;
+		return static_cast<float>(TimeDelta / TickResolution) * PixelsPerSecond;
 	}
 
 	/**
